Renamed TLCTripYellowTaxi Class to TLC::YellowTaxiTrip Class
