+++ lib/datasets.rb
@@ -29,7 +29,7 @@ require_relative "datasets/quora-duplicate-question-pair"
 require_relative "datasets/rdataset"
 require_relative "datasets/seaborn"
 require_relative "datasets/sudachi-synonym-dictionary"
-require_relative "datasets/tlc-trip-yellow-taxi"
+require_relative "datasets/tlc/yellow-taxi-trip"
 require_relative "datasets/wikipedia"
 require_relative "datasets/wikipedia-kyoto-japanese-english"
 require_relative "datasets/wine"
